tweaks to AssignmentNode (convertToJott)
